filter free vars in lambda
