+++ core/src/test/scala/org/bykn/bosatsu/TestUtils.scala
@@ -1,8 +1,12 @@
 package org.bykn.bosatsu
 
 import org.bykn.bosatsu.rankn._
+import cats.data.Validated
+import org.scalatest.{Assertion, Assertions}
 
 import fastparse.all.Parsed
+
+import Assertions.fail
 
 object TestUtils {
   import TestParseUtils.region
@@ -40,4 +44,16 @@ object TestUtils {
         sys.error(s"failed to parse: $str: $exp at $idx in region ${region(str, idx)} with trace: ${extra.traced.trace}")
     }
   }
+
+  def checkLast(statement: String)(fn: TypedExpr[Declaration] => Assertion): Assertion =
+    Statement.parser.parse(statement) match {
+      case Parsed.Success(stmt, _) =>
+        Package.inferBody(PackageName.parts("Test"), Nil, stmt) match {
+          case Validated.Invalid(errs) => fail(errs.toList.map(_.message(Map.empty)).mkString("\n"))
+          case Validated.Valid((_, lets)) =>
+            fn(lets.last._3)
+        }
+      case Parsed.Failure(exp, idx, extra) =>
+        fail(s"failed to parse: $statement: $exp at $idx with trace: ${extra.traced.trace}")
+    }
 }
